requirements.txt
PyYAML~=6.0.1
angr~=9.2.100
scikit-learn~=1.3.2
memory-profiler~=0.61.0
archinfo~=9.2.100
claripy~=9.2.100
pandas~=2.0.3
scapy~=2.5.0
Netzob~=2.0.0
ipython~=8.12.3
tracer~=0.1
pyvex~=9.2.100
thefuzz~=0.22.1
cle~=9.2.100
capstone~=5.0.0.post1